requirements.txt
PySide6
Pillow
requests
keyring
google-genai
pytest
pytest-cov
pytest-qt
